Refactor classes into correct folder
