+++ src/main/java/luoyi/duke/commands/DeadlineCommand.java
@@ -5,22 +5,22 @@ import luoyi.duke.common.TextFormatter;
 import luoyi.duke.data.Duke;
 import luoyi.duke.data.IDuke;
 import luoyi.duke.data.exception.DukeIllegalArgumentException;
-import luoyi.duke.data.task.Event;
+import luoyi.duke.data.task.Deadline;
 import luoyi.duke.data.task.ITask;
 import luoyi.duke.data.task.TaskList;
 import luoyi.duke.storage.Storage;
 
-public class EventCommand extends Command {
+public class DeadlineCommand extends Command {
     private final String description;
     private final String time;
-    private EventCommand(String description, String time, IDuke duke) {
+    private DeadlineCommand(String description, String time, IDuke duke) {
         super(-1, duke);
         this.description = description;
         this.time = time;
     }
 
-    public static EventCommand getEventCommand(String description, String time) {
-        return new EventCommand(description, time, null);
+    public static DeadlineCommand getDeadlineCommand(String description, String time) {
+        return new DeadlineCommand(description, time, null);
     }
 
     @Override
@@ -28,22 +28,22 @@ public class EventCommand extends Command {
         if (duke == null) {
             throw new RuntimeException(Message.ERR_DUKE_NOT_INIT.toString());
         }
-        return handleEvent(description, time);
+        return handleDeadline(description, time);
     }
 
-    private IDuke handleEvent(String description, String time)
+    private IDuke handleDeadline(String description, String time)
             throws DukeIllegalArgumentException {
         if (description.matches("\\s*")) {
             throw new DukeIllegalArgumentException(
-                    "The description of event cannot be empty!");
+                    "The description of deadline cannot be empty!");
         }
         if (time.matches("\\s*")) {
             throw new DukeIllegalArgumentException(
-                    "The time of event cannot be empty!");
+                    "The time of deadline cannot be empty!");
         }
-        ITask task = Event.getEvent(description, time);
+        ITask task = Deadline.getDeadline(description, time);
         IDuke newDuke = storeTask(task);
-        System.out.println(TextFormatter.getFormattedText(
+        System.out.print(TextFormatter.getFormattedText(
                 "Got it. I've added this task:\n\t" + task.toString()
                         + "\nNow you have " +  newDuke.getNumTask()
                         + " task(s) in the list."));
@@ -60,6 +60,6 @@ public class EventCommand extends Command {
 
     @Override
     public Command setDuke(IDuke duke) {
-        return new EventCommand(description, time, duke);
+        return new DeadlineCommand(description, time, duke);
     }
 }
