+++ src/main/java/luoyi/duke/commands/EventCommand.java
@@ -5,20 +5,22 @@ import luoyi.duke.common.TextFormatter;
 import luoyi.duke.data.Duke;
 import luoyi.duke.data.IDuke;
 import luoyi.duke.data.exception.DukeIllegalArgumentException;
+import luoyi.duke.data.task.Event;
 import luoyi.duke.data.task.ITask;
 import luoyi.duke.data.task.TaskList;
-import luoyi.duke.data.task.ToDo;
 import luoyi.duke.storage.Storage;
 
-public class ToDoCommand extends Command {
+public class EventCommand extends Command {
     private final String description;
-    private ToDoCommand(String description, IDuke duke) {
+    private final String time;
+    private EventCommand(String description, String time, IDuke duke) {
         super(-1, duke);
         this.description = description;
+        this.time = time;
     }
 
-    public static ToDoCommand getToDoCommand(String description) {
-        return new ToDoCommand(description, null);
+    public static EventCommand getEventCommand(String description, String time) {
+        return new EventCommand(description, time, null);
     }
 
     @Override
@@ -26,20 +28,25 @@ public class ToDoCommand extends Command {
         if (duke == null) {
             throw new RuntimeException(Message.ERR_DUKE_NOT_INIT.toString());
         }
-        return handleToDo(description);
+        return handleEvent(description, time);
     }
 
-    private IDuke handleToDo(String description) throws DukeIllegalArgumentException {
+    private IDuke handleEvent(String description, String time)
+            throws DukeIllegalArgumentException {
         if (description.matches("\\s*")) {
             throw new DukeIllegalArgumentException(
-                    "The description of todo cannot be empty!");
+                    "The description of event cannot be empty!");
         }
-        ITask task = ToDo.getToDo(description);
+        if (time.matches("\\s*")) {
+            throw new DukeIllegalArgumentException(
+                    "The time of event cannot be empty!");
+        }
+        ITask task = Event.getEvent(description, time);
         IDuke newDuke = storeTask(task);
-        System.out.println(TextFormatter.getFormattedText(
+        System.out.print(TextFormatter.getFormattedText(
                 "Got it. I've added this task:\n\t" + task.toString()
-                        + "\nNow you have "
-                        +  newDuke.getNumTask() + " task(s) in the list."));
+                        + "\nNow you have " +  newDuke.getNumTask()
+                        + " task(s) in the list."));
         return newDuke;
     }
 
@@ -53,6 +60,6 @@ public class ToDoCommand extends Command {
 
     @Override
     public Command setDuke(IDuke duke) {
-        return new ToDoCommand(description, duke);
+        return new EventCommand(description, time, duke);
     }
 }
