+++ src/main/java/luoyi/duke/common/Message.java
