+++ src/main/java/luoyi/duke/data/IDuke.java
